refactor: remove `api-client` functions
