+++ server/api/brevia/collections.get.ts
@@ -1,12 +1,13 @@
 export default defineEventHandler(async (event) => {
+  const config = useRuntimeConfig(event);
+  let url = config.apiBaseUrl + `/collections`;
   const query = getQuery(event);
-  let url = '/collections';
   if (query.uuid) {
     url += `/${query.uuid}`;
     delete query.uuid;
   }
   try {
-    const response: any = await apiFetch(url, { query }, event);
+    const response: any = await $fetch(url, { query }, event);
 
     if ('name' in query && response.length > 0) {
       return response[0];
